refactor: move pbias to hydro_impl
