+++ src/scores/continuous/__init__.py
@@ -12,7 +12,7 @@ from scores.continuous.flip_flop_impl import (
     flip_flop_index,
     flip_flop_index_proportion_exceeding,
 )
-from scores.continuous.hydro_impl import kge, nse
+from scores.continuous.hydro_impl import kge, nse, pbias
 from scores.continuous.interval_impl import interval_score, quantile_interval_score
 from scores.continuous.quantile_loss_impl import quantile_score
 from scores.continuous.standard_impl import (
@@ -21,7 +21,6 @@ from scores.continuous.standard_impl import (
     mean_error,
     mse,
     multiplicative_bias,
-    pbias,
     percent_within_x,
     rmse,
 )
